[update]cart_item logic repair(error check)

+++ app/controllers/public/cart_items_controller.rb
@@ -10,12 +10,14 @@ class Public::CartItemsController < Public::ApplicationController
 
   def update
     @cart_item = CartItem.find(params[:id])
-    if @cart_item.update(cart_item_update_params)
-      redirect_to request.referer, notice: "カート内の商品(#{@cart_item.item.name})の数量を変更しました。"
+    if (params[:cart_item][:amount]).to_i <= 10
+      if @cart_item.update(cart_item_update_params)
+        redirect_to request.referer, notice: "カート内の商品(#{@cart_item.item.name})の数量を変更しました。"
+      else
+        redirect_to request.referer, notice: "数量の更新が反映されませんでした。再度お試しください。"
+      end
     else
-      @cart_items = current_customer.cart_items
-      @total_payment_no_shipfee = CartItem.total_payment_no_shipfee(@cart_items)
-      render "index"
+      redirect_to request.referer, notice: "数量は10個以内でご指定ください。"
     end
   end
 
@@ -24,18 +26,20 @@ class Public::CartItemsController < Public::ApplicationController
     if cart_item.destroy
       redirect_to request.referer, notice: "カート内の商品を1件(#{cart_item.item.name})削除しました。"
     else
-      @cart_items = current_customer.cart_items
-      @total_payment_no_shipfee = CartItem.total_payment_no_shipfee(@cart_items)
-      render "index"
+      redirect_to request.referer, notice: "削除が正常に完了しませんでした。再度お試しください。"
     end
   end
 
   def destroy_all
     @cart_items = current_customer.cart_items
-    @cart_items.each do |cart_item|
-      cart_item.destroy
+    if @cart_items
+      @cart_items.each do |cart_item|
+        cart_item.destroy
+      end
+      redirect_to request.referer, notice: "カート内の商品は全て削除されました。"
+    else
+      redirect_to request.referer, notice: "削除が正常に完了しませんでした。再度お試しください。"
     end
-    redirect_to request.referer, notice: "カート内の商品は全て削除されました。"
   end
 
   private
